Move column related logic under ColumnAnnotation namespace
