+++ lib/annotate_rb/model_annotator.rb
@@ -17,10 +17,7 @@ module AnnotateRb
     autoload :ModelFileAnnotator, 'annotate_rb/model_annotator/model_file_annotator'
     autoload :ModelWrapper, 'annotate_rb/model_annotator/model_wrapper'
     autoload :AnnotationGenerator, 'annotate_rb/model_annotator/annotation_generator'
-    autoload :ColumnAttributesBuilder, 'annotate_rb/model_annotator/column_attributes_builder'
-    autoload :ColumnTypeBuilder, 'annotate_rb/model_annotator/column_type_builder'
-    autoload :ColumnWrapper, 'annotate_rb/model_annotator/column_wrapper'
-    autoload :ColumnAnnotationBuilder, 'annotate_rb/model_annotator/column_annotation_builder'
+    autoload :ColumnAnnotation, 'annotate_rb/model_annotator/column_annotation'
     autoload :IndexAnnotationBuilder, 'annotate_rb/model_annotator/index_annotation_builder'
     autoload :ForeignKeyAnnotationBuilder, 'annotate_rb/model_annotator/foreign_key_annotation_builder'
     autoload :RelatedFilesListBuilder, 'annotate_rb/model_annotator/related_files_list_builder'
