+++ lib/annotate_rb/model_annotator/column_annotation.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module ColumnAnnotation
+      autoload :AttributesBuilder, 'annotate_rb/model_annotator/column_annotation/attributes_builder'
+      autoload :TypeBuilder, 'annotate_rb/model_annotator/column_annotation/type_builder'
+      autoload :ColumnWrapper, 'annotate_rb/model_annotator/column_annotation/column_wrapper'
+      autoload :AnnotationBuilder, 'annotate_rb/model_annotator/column_annotation/annotation_builder'
+    end
+  end
+end
