+++ spec/lib/annotate_rb/model_annotator/column_annotation/attributes_builder_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe AnnotateRb::ModelAnnotator::ColumnAttributesBuilder do
+RSpec.describe AnnotateRb::ModelAnnotator::ColumnAnnotation::AttributesBuilder do
   include AnnotateTestHelpers
 
   describe '#build' do
